Restructure tests folder
