+++ src/__tests__/components/NavigationBar.test.js
@@ -1,6 +1,6 @@
 import { BrowserRouter } from 'react-router-dom';
 import renderer from 'react-test-renderer';
-import NavigationBar from '../NavigationBar';
+import NavigationBar from '../../components/NavigationBar';
 
 test('NavigationBar renders correctly', () => {
   const tree = renderer
