+++ src/__tests__/components/TextBody.test.js
@@ -1,5 +1,5 @@
 import renderer from 'react-test-renderer';
-import TextBody from '../TextBody';
+import TextBody from '../../components/TextBody';
 
 test('TextBody renders correctly', () => {
   const paragraphs = ['Paragraph 1', 'Paragraph 2', 'Paragraph 3'];
