+++ src/__tests__/components/WebsiteTitle.test.js
@@ -1,5 +1,5 @@
 import renderer from 'react-test-renderer';
-import WebsiteTitle from '../WebsiteTitle';
+import WebsiteTitle from '../../components/WebsiteTitle';
 
 test('WebsiteTitle renders correctly', () => {
   const tree = renderer
